backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
daphne==4.0.0
pytesseract==0.3.10
zope-interface>=5.4.0
Twisted>=22.10.0
requests>=2.31.0
gunicorn==21.2.0
uvicorn[standard]==0.24.0.post1
fastapi==0.104.1

